# backend/docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PRM Marketplace API",
    "description" : "Backend API for the student/resident marketplace",
    "contact" : {
      "name" : "PRM Marketplace Team"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/cart-items/{id}/quantity" : {
      "put" : {
        "tags" : [ "cart-item-controller" ],
        "operationId" : "updateQuantity",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "quantity",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CartItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/bulletin-posts/{id}" : {
      "get" : {
        "tags" : [ "bulletin-post-controller" ],
        "operationId" : "read_6",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BulletinPost"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "bulletin-post-controller" ],
        "operationId" : "update_8",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulletinPost"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BulletinPost"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "bulletin-post-controller" ],
        "operationId" : "delete_3",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/orders/checkout" : {
      "post" : {
        "tags" : [ "order-controller" ],
        "operationId" : "checkout",
        "parameters" : [ {
          "name" : "shippingAddressId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cart-items" : {
      "get" : {
        "tags" : [ "cart-item-controller" ],
        "operationId" : "getAll_7",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CartItem"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "cart-item-controller" ],
        "operationId" : "addToCart",
        "parameters" : [ {
          "name" : "productId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "quantity",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CartItem"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "cart-item-controller" ],
        "operationId" : "clearCart",
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/bulletin-posts" : {
      "get" : {
        "tags" : [ "bulletin-post-controller" ],
        "operationId" : "getAll_8",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BulletinPost"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "bulletin-post-controller" ],
        "operationId" : "create_9",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BulletinPost"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BulletinPost"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/post-likes/post/{postId}" : {
      "get" : {
        "tags" : [ "post-like-controller" ],
        "operationId" : "hasLiked",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comments/post/{postId}" : {
      "get" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "getByPost",
        "parameters" : [ {
          "name" : "postId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comments/author/{authorId}" : {
      "get" : {
        "tags" : [ "comment-controller" ],
        "operationId" : "getByAuthor",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cart-items/{id}" : {
      "get" : {
        "tags" : [ "cart-item-controller" ],
        "operationId" : "read_16",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CartItem"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "cart-item-controller" ],
        "operationId" : "delete_11",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/bulletin-posts/author/{authorId}" : {
      "get" : {
        "tags" : [ "bulletin-post-controller" ],
        "operationId" : "getByAuthor_1",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BulletinPost"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "STUDENT", "FACULTY", "VENDOR", "RESIDENT" ],
            "readOnly" : true
          },
          "phone" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "verified" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "avatarUrl" : {
            "type" : "string"
          }
        }
      },
...
      "VendorProfile" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "readOnly" : true
          },
          "businessName" : {
            "type" : "string"
          },
          "registrationNo" : {
            "type" : "string"
          },
          "verified" : {
            "type" : "boolean",
            "readOnly" : true
          },
          "ratingAvg" : {
            "type" : "number",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
...
      "Product" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "price" : {
            "type" : "number"
          },
          "stockQuantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "category" : {
            "type" : "string"
          },
          "imageUrl" : {
            "type" : "string"
          },
          "condition" : {
            "type" : "string",
            "enum" : [ "NEW", "LIKE_NEW", "GOOD", "FAIR", "POOR" ]
          },
          "city" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "vendor" : {
            "$ref" : "#/components/schemas/VendorProfile",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "active" : {
            "type" : "boolean"
          }
        }
      },
...
      "Address" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "line1" : {
            "type" : "string"
          },
          "line2" : {
            "type" : "string"
          },
          "suburb" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "province" : {
            "type" : "string"
          },
          "postalCode" : {
            "type" : "string"
          },
          "country" : {
            "type" : "string"
          },
          "defaultAddress" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "singleLine" : {
            "type" : "string"
          }
        }
      },
      "Order" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "buyer" : {
            "$ref" : "#/components/schemas/User",
            "readOnly" : true
          },
          "status" : {
            "type" : "string",
            "enum" : [ "PENDING", "CONFIRMED", "SHIPPED", "DELIVERED", "CANCELLED", "REFUNDED" ],
            "readOnly" : true
          },
          "totalAmount" : {
            "type" : "number",
            "readOnly" : true
          },
          "shippingAddress" : {
            "$ref" : "#/components/schemas/Address"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
      "BulletinPost" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "author" : {
            "$ref" : "#/components/schemas/User",
            "readOnly" : true
          },
          "title" : {
            "type" : "string"
          },
          "body" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          },
          "imageUrl" : {
            "type" : "string"
          },
          "commentCount" : {
            "type" : "integer",
            "format" : "int32",
            "readOnly" : true
          },
          "likeCount" : {
            "type" : "integer",
            "format" : "int32",
            "readOnly" : true
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
      "Comment" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "post" : {
            "$ref" : "#/components/schemas/BulletinPost",
            "writeOnly" : true
          },
          "author" : {
            "$ref" : "#/components/schemas/User",
            "readOnly" : true
          },
          "parent" : {
            "$ref" : "#/components/schemas/Comment",
            "writeOnly" : true
          },
          "body" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          },
          "reply" : {
            "type" : "boolean"
          }
        }
      },
      "CartItem" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "user" : {
            "$ref" : "#/components/schemas/User",
            "readOnly" : true
          },
          "product" : {
            "$ref" : "#/components/schemas/Product"
          },
          "quantity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "addedAt" : {
            "type" : "string",
            "format" : "date-time",
            "readOnly" : true
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}